+++ __init__.py
@@ -1,5 +1,5 @@
 """
 Oktopios — Un langage de programmation moderne et expressif 🐙
 """
-__version__ = "0.2.8"
+__version__ = "0.3.0"
 __author__ = "Mouanwiya Ali Soule"